fixed dome shit

+++ libft.h
@@ -56,7 +56,7 @@ void				ft_putnbr_fd(int n, int fd);
 char				*ft_strjoin(char const *s1, char const *s2);
 char				*ft_strtrim(char const *s1, char const *set);
 char				**ft_split(char const *s, char c);
-char				*ft_itoa(long n);
+char				*ft_itoa(unsigned long long n);
 char				*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 t_list				*ft_lstnew(void *content);
 void				ft_lstadd_front(t_list **alst, t_list *new);
